chore: bump version to v2.1.0-beta
Extension host, math toggle, content-only zoom, live server type fixes,
editor transform setMarkdown fallback. Refresh attributions.

+++ src-tauri/src/m045_ext/live_server.rs
@@ -158,8 +158,8 @@ pub struct ExtContextResponse {
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct CursorCoords {
-    pub line: u32,
-    pub ch: u32,
+    pub line: f64,
+    pub ch: f64,
 }
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
